some refactoring old camera code and add more checks

See merge request ic/mobile-app!13

+++ src/scenes/PhotoScene.js
@@ -3,8 +3,6 @@ import PropTypes from 'prop-types'
 import {connect} from 'react-redux'
 import {
     ActivityIndicator,
-    ImageBackground,
-    Modal,
     StyleSheet,
     View,
     Dimensions,
@@ -13,8 +11,6 @@ import {
     Text,
     Platform,
     Animated,
-    Easing,
-    TouchableWithoutFeedback,
     Vibration,
     SafeAreaView,
     Alert,
@@ -23,17 +19,16 @@ import {
 import {addPhoto, uploadPhoto, clearPhoto, setCameraType, deleteImage} from '../actions/photo'
 import {getFileSize, getPhotoPath, isIphoneX} from '../utils/util'
 import {goToSettings, back} from '../actions/navigation'
-import {Container, Header, Left, Right, Icon, Title, ListItem, Button, Footer} from 'native-base'
+import {Container} from 'native-base'
 import I18n from 'react-native-i18n'
 import {NavigationActions} from 'react-navigation'
 import _ from 'lodash'
 import {RNCamera} from 'react-native-camera';
 import {backIcon, cameraButton, closeIcon, flashAuto, flashOff, flashOn} from "../utils/images";
 import {copyFile, unlink, stat} from 'react-native-fs';
-import {decorator as sensors} from "react-native-sensors";
 import ImageResizer from 'react-native-image-resizer';
 import Orientation from "react-native-orientation";
-import {PinchGestureHandler} from "react-native-gesture-handler"
+//import {PinchGestureHandler} from "react-native-gesture-handler"
 import {Map} from "immutable";
 import {basename} from "react-native-path";
 //import ARCamera from "ar-shelf-camera"
@@ -77,12 +72,6 @@ class PhotoScene extends Component {
         };
     }
 
-    prepareRatio = async () => {
-        if (Platform.OS === 'android' && this.camera) {
-            const ratios = await this.camera.getSupportedRatiosAsync();
-        }
-    };
-
     fadeCamera() {
         Animated.timing(
             this.state.fadeValue,
@@ -104,10 +93,8 @@ class PhotoScene extends Component {
     componentDidMount() {
         StatusBar.setHidden(true);
         const photoIndex = this.props.navigation.getParam("photoIndex", null);
-        this.setState({photoIndex})
-
+        this.setState({photoIndex});
     }
-
 
     onBackPress = () => {
         this.props.navigation.dispatch(NavigationActions.back());
@@ -116,9 +103,6 @@ class PhotoScene extends Component {
     componentWillUnmount() {
         StatusBar.setHidden(false);
         const {backHandler} = this.props.navigation.state.params;
-        if (this.ifAR() && this.camera !== null) {
-            this.camera.stopSession();
-        }
         backHandler && backHandler(this.state.finalUri);
     }
 
@@ -130,41 +114,6 @@ class PhotoScene extends Component {
         this.props.dispatch(NavigationActions.back())
     };
 
-    async changeLayout() {
-    }
-
-    getRotate() {
-        const {Accelerometer} = this.props;
-        let x = 0;
-        let y = 0;
-        let z = 0;
-        if (Accelerometer) {
-            x = Accelerometer.x;
-            y = Accelerometer.y;
-            z = Accelerometer.z;
-        }
-
-        let rotate = 0;
-
-        if (Platform.OS === 'ios') {
-            if (x > 0.60 && z > -0.80 && Math.abs(y) < 0.60) {
-                rotate = 90
-            }
-            if (x < -0.60 && z > -0.80 && Math.abs(y) < 0.60) {
-                rotate = -90
-            }
-        } else {
-            if (x > 6 && z < 8 && y < 6) {
-                rotate = -90
-            }
-            if (x < -6 && z < 8 && y < 6) {
-                rotate = 90
-            }
-        }
-
-        return rotate;
-    }
-
     changeFlashMode = () => {
         if (this.state.flashMode === RNCamera.Constants.FlashMode.off) {
             this.setState({flashMode: RNCamera.Constants.FlashMode.on})
@@ -172,58 +121,9 @@ class PhotoScene extends Component {
         if (this.state.flashMode === RNCamera.Constants.FlashMode.on) {
             this.setState({flashMode: RNCamera.Constants.FlashMode.off})
         }
-        if (this.ifAR()) {
-            this.camera.toggleFlash();
-        }
-    };
-
-    getOrientation() {
-        return new Promise((resolve, reject) => {
-            Orientation.getOrientation((err, orientation) => {
-                return (err) ? reject(err) : resolve(orientation);
-            });
-        })
-    }
-
-    confirmPlane() {
-        if (this.camera && this.state.planeCaptured === true) {
-            this.camera.confirmPlane();
-        }
-    };
-
-    renderAr = () => {
-        return <TouchableOpacity onPress={() => this.confirmPlane()}><ARCamera
-            ref={ref => {
-                if (this.startedSession === false && ref !== null) {
-                    ref.startSession();
-                    this.startedSession = true;
-                }
-                this.camera = ref;
-            }}
-            flashMode={this.state.flashMode}
-            width={1500}
-            quality={0.9}
-            onCameraReady={() => {
-                this.setState({arReady: true});
-            }}
-            onPlaneCaptureStarted={() => {
-                console.log("onPlaneCaptureStarted")
-            }}
-            onPlaneCaptureSuccess={() => {
-                this.setState({planeCaptured: true});
-                Vibration.vibrate(DURATION)
-            }}
-            onPlaneCaptureLost={() => {
-                this.setState({planeCaptured: false});
-            }}
-        /></TouchableOpacity>
     };
 
     renderCameraElement = () => {
-        const props = {};
-        if (this.state.ratio) {
-            props.ratio = "4:3"
-        }
         return <RNCamera
             ref={ref => {
                 this.camera = ref;
@@ -232,7 +132,6 @@ class PhotoScene extends Component {
             style={[styles.preview]}
             type={RNCamera.Constants.Type.back}
             flashMode={this.state.flashMode}
-            onCameraReady={this.prepareRatio}
             ratio="4:3"
         />
     };
@@ -247,18 +146,7 @@ class PhotoScene extends Component {
     }
 
     ifAR = () => {
-        //Disable AR
         return false;
-        const {pins, pin} = this.props;
-        const enable_ar_camera = pins[pin].enable_ar_camera;
-        return Platform.OS === 'ios' && Platform.Version >= "11.3" && this.state.cameraTypeIos === AR_CAMERA && enable_ar_camera === true
-    };
-
-    ifIosStandard = () => {
-        return true;
-        const {pins, pin} = this.props;
-        const enable_ar_camera = pins[pin].enable_ar_camera;
-        return Platform.OS === 'ios' && (this.state.cameraTypeIos === STANDARD_CAMERA || enable_ar_camera !== true)
     };
 
     takePicture = async () => {
@@ -275,53 +163,51 @@ class PhotoScene extends Component {
         this.setState({processPhoto: true});
 
         const options = {quality: 0.9, exif: true, width: 1500, skipProcessing: true};
-        let rotate = 0;
-
-        if (Platform.OS === "android") {
-            rotate = this.getRotate();
-        }
-
-        const initial = await this.getOrientation();
 
         let data;
-        if (this.ifAR()) {
-            try {
-                data = await this.camera.takePicture(1500, 0.9);
-            } catch (error) {
-                data = null;
-            }
-        } else {
+        try {
             data = await this.camera.takePictureAsync(options);
+        } catch (error) {
+            data = null;
+            this.takePictureStatus = false;
+            this.setState({processPhoto: false});
+            return Alert.alert("Фото не сделано", _.get(error, "message", ""));
         }
 
         if (data === null) {
             this.takePictureStatus = false;
             this.setState({processPhoto: false});
-            return alert("Фото не сделано");
-        }
-
-        if (this.ifIosStandard() && initial !== 'PORTRAIT') {
-            rotate = 0
-        }
-
-        if (Platform.OS !== 'ios' && data.width > data.height) {
-            rotate = rotate + 90;
-        }
-
-        if (Platform.OS === 'ios' && initial === 'PORTRAIT' && rotate !== 0) {
-            rotate = 0
+            return Alert.alert("Фото не сделано");
         }
 
         const width = (Platform.OS === 'ios') ? data.width : 1500;
         const height = (Platform.OS === 'ios') ? data.height : 1500;
 
-        const result = await ImageResizer.createResizedImage(data.uri, width, height, 'JPEG', 90);
+        let result;
+        try {
+            result = await ImageResizer.createResizedImage(data.uri, width, height, 'JPEG', 90);
+        } catch (error) {
+            this.takePictureStatus = false;
+            this.setState({processPhoto: false});
+            return Alert.alert("Ошибка обработки", _.get(error, "message", ""));
+        }
+
+        if (!result) {
+            this.takePictureStatus = false;
+            this.setState({processPhoto: false});
+            return Alert.alert("Нет результата обработки");
+        }
+
         if (result.size === 0) {
             this.takePictureStatus = false;
             this.setState({processPhoto: false});
-            await unlink(data.uri);
-            await unlink(result.uri);
-            return alert("Размер фото 0 байт, фото будет удаленно");
+            try {
+                await unlink(data.uri);
+                await unlink(result.uri);
+            } catch (error) {
+                return Alert.alert("Ошибка удаления пустых файлов", _.get(error, "message", ""))
+            }
+            return Alert.alert("Размер фото 0 байт, фото будет удаленно");
         }
 
         this.setState(state => {
@@ -331,16 +217,34 @@ class PhotoScene extends Component {
         });
 
         const index = photoIndex ? photoIndex : this.getPhotosIndex();
-        await unlink(data.uri);
+        try {
+            await unlink(data.uri);
+        } catch (error) {
+            this.takePictureStatus = false;
+            this.setState({processPhoto: false});
+            return Alert.alert("Ошибка удаления", _.get(error, "message", ""))
+        }
+
         const path = getPhotoPath(result.path);
-        await copyFile(result.uri, path);
-        await unlink(result.uri);
+        try {
+            await copyFile(result.uri, path);
+        } catch (error) {
+            this.takePictureStatus = false;
+            this.setState({processPhoto: false});
+            return Alert.alert("Ошибка копирования", _.get(error, "message", ""))
+        }
+
+        try {
+            await unlink(result.uri);
+        } catch (error) {
+            this.takePictureStatus = false;
+            this.setState({processPhoto: false});
+            return Alert.alert("Ошибка удаления результата", _.get(error, "message", ""))
+        }
+
         const finalUri = "file://" + path;
         const {id} = this.props.navigation.state.params;
         await this.props.addPhoto(finalUri, id, index);
-        if (Map(this.props.visits.entities.offline).count() === 0) {
-            //this.props.uploadPhoto(finalUri, id, null, undefined, index);
-        }
 
         if (photoUUID) {
             try {
@@ -357,17 +261,20 @@ class PhotoScene extends Component {
             }
         }
 
-        this.setState(state => (
-            {photoCount: state.photoCount + 1, processPhoto: false, finalUri}
-        ), async () => {
-            if (photoUUID === null) {
-                this.takePictureStatus = false;
-            } else if (photoUUID && this.state.photoCount < 1) {
-                this.takePictureStatus = false;
-            } else if (photoUUID && this.state.photoCount === 1) {
-                this.onBackPress();
-            }
-        });
+        setTimeout(() => {
+            this.setState(state => (
+                {photoCount: state.photoCount + 1, processPhoto: false, finalUri}
+            ), () => {
+                if (photoUUID === null) {
+                    this.takePictureStatus = false;
+                } else if (photoUUID && this.state.photoCount < 1) {
+                    this.takePictureStatus = false;
+                } else if (photoUUID && this.state.photoCount === 1) {
+                    this.onBackPress();
+                }
+            });
+        }, 100)
+
     };
 
     changeZoom = _.throttle(() => {
@@ -393,79 +300,30 @@ class PhotoScene extends Component {
     };
 
     selectCamera() {
-        let camera = null;
-        if (Platform.OS === 'ios') {
-            camera = (this.ifAR())
-                ? this.renderAr()
-                : this.renderCameraElement();
-        } else {
-            camera = this.renderCameraElement();
-        }
-
-        return camera
+        return this.renderCameraElement()
     }
-
-    closeArCamera = () => {
-        this.setState({cameraTypeIos: STANDARD_CAMERA});
-        if (this.camera) {
-            this.camera.stopSession();
-        }
-    };
 
     renderBottomControl = () => {
         const overButton = (this.state.processPhoto === true) ?
             <View style={{position: "absolute", zIndex: 9}}><ActivityIndicator/></View> :
             (this.state.photoCount > 0) ? <Text style={styles.photoCount}>{this.state.photoCount}</Text> : null;
         const buttonTint = (this.state.processPhoto === true) ? {tintColor: "#555"} : {tintColor: "white"};
-
-        const text = !this.state.planeCaptured ? I18n.t("arCamera.captureStarted")
-            : I18n.t("arCamera.captureSuccess");
-
-        if (this.ifAR()) {
-            if (this.state.arReady) {
-                return (
-                    <TouchableOpacity onPress={this.takePicture}>
-                        {overButton}
-                        <Image style={[styles.cameraImage, buttonTint]} source={cameraButton}/>
-                    </TouchableOpacity>
-                )
-            } else {
-                return (
-                    <View style={{alignItems: "center"}}>
-                        <Text style={styles.arText}>{text}</Text>;
-                        <TouchableOpacity
-                            onPress={this.closeArCamera}
-                            style={styles.cancelButton}>
-                            <Text style={styles.cancelAr}>Отмена</Text>
-                        </TouchableOpacity>
-                    </View>
-                )
-            }
-
-        } else {
-            return (
-                <TouchableOpacity onPress={this.takePicture} style={styles.capture}>
-                    {overButton}
-                    <Image style={[styles.cameraImage, buttonTint]} source={cameraButton}/>
-                </TouchableOpacity>
-            )
-        }
+        return (
+            <TouchableOpacity onPress={this.takePicture} style={styles.capture}>
+                {overButton}
+                <Image style={[styles.cameraImage, buttonTint]} source={cameraButton}/>
+            </TouchableOpacity>
+        )
     };
 
     renderCamera() {
-        const topPadding = (Platform.OS === "ios") ? 0 : 15;
-        const rotate = this.getRotate();
-
         let flashIcon = flashAuto;
-
         if (this.state.flashMode === RNCamera.Constants.FlashMode.on) {
             flashIcon = flashOn;
         }
-
         if (this.state.flashMode === RNCamera.Constants.FlashMode.off) {
             flashIcon = flashOff;
         }
-        const {id} = this.props.navigation.state.params;
         const style = (!this.ifAR()) ? {flex: 1} : {};
         const diff = height - heightCamera;
         return (
@@ -476,7 +334,7 @@ class PhotoScene extends Component {
                     <TouchableOpacity onPress={this.onBackPress} style={{paddingLeft: 16}}
                                       hitSlop={{top: 50, left: 50, bottom: 50, right: 50}}>
                         <Image source={closeIcon}
-                               style={{tintColor: "white", transform: [{rotate: -1 * rotate + 'deg'}]}}/>
+                               style={{tintColor: "white"}}/>
                     </TouchableOpacity>
                 </SafeAreaView>
                 <SafeAreaView style={{position: "absolute", right: 0, top: 22}}>
@@ -485,7 +343,6 @@ class PhotoScene extends Component {
                         <Image source={flashIcon} style={{tintColor: "white", width: 24, height: 24}}/>
                     </TouchableOpacity>
                 </SafeAreaView>
-
                 <View style={diff - 55 > 65 ? styles.bottomRowFlex : styles.bottomRowAbs}>
                     {this.renderBottomControl()}
                 </View>
@@ -528,7 +385,7 @@ export default connect((state) => {
         }
     },
     {uploadPhoto, addPhoto, back, clearPhoto, setCameraType, deleteImage}
-)((Platform.OS === "android") ? sensors({Accelerometer: {updateInterval: 300}})(PhotoScene) : PhotoScene)
+)(PhotoScene)
 
 
 const styles = StyleSheet.create({
